Fixes terrain tile placement algo for certain configs.
In the case where a terrain tile overflows the bottom of the
terrain object's bounding box, a mask is now always applied -
this is to fix a bug in the math that sometimes caused underflow
(positioning the tiles too high so that they didn't fill up the
bounding box). In exchange it is a bit more liberal with masking,
which may come into play when doing animated worlds.