Adds growing level with terrain spawning on pre-set days.
You can now use the level editor to set the 'day spawn' field, which
determines on which day the piece of terrain spawns. So far things like
decorations and spawn points are not affected by the 'day spawn', but
they will need to be in the future so that for instance a tree and
gravestone can spawn only once the platform below them has spawned.

Additionally collision with the player does not work very well right
now when the terrain is spawning - there will need to be a better way
of 'pushing' objects out of the way as terrain spawns upward. Perhaps
something like the following:

1) Check if objects need to change size due to spawning - if so, have
   them check for objects in the way and push them up as well.
2) Have non-terrain objects update their positions, but have them check
   if this would collide with terrain beforehand and only move as far
   as terrain allows. No more moving into the terrain and then having
   the terrain figure out how to push the object out.
3) Update other animations/timers/handle input for both non-terrain
   and terrain objects (i.e. do updatePreCollision()).
4) Do collision detection and response.
5) Do updatePostCollision() for both non-terrain and terrain objects.

One more thing is that right now the new terrain pops out of the
ground after the sun comes up - I might want to change the exact time
that terrain starts to spawn, and maybe the view should also change
while terrain is spawning (i.e. zoom out so we can see the whole map,
pop out new terrain, then zoom back in to the player).